- Load pending polls in click - cleanup redundant and unused code
