+++ src/StratisMasternodeDashboard/Models/SidechainNodeModel.cs
@@ -6,10 +6,7 @@ namespace Stratis.FederatedSidechains.AdminDashboard.Models
     public class SidechainNodeModel : StratisNodeModel
     {
         public List<PendingPoll> PoAPendingPolls { get; set; }
-        public int FederationMemberCount { get; set; }
-        public int BlockProducerHits { get; set; }
-        public decimal BlockProducerHitsValue { get; set; }
+        public int FederationMemberCount { get; set; }       
         public List<PendingPoll> KickFederationMemberPolls { get; set; }
-        public string SidechainMiningAddress { get; set; }
     }
 }
