+++ src/StratisMasternodeDashboard/Services/MasterNodeDataService.cs
@@ -30,9 +30,7 @@ namespace Stratis.FederatedSidechains.AdminDashboard.Services
         {
             await base.Update().ConfigureAwait(false);
 
-            PendingPolls = await UpdatePolls().ConfigureAwait(false);
             KickFederationMememberPendingPolls = await UpdateKickFederationMemberPolls().ConfigureAwait(false);
-            FederationMemberCount = await UpdateFederationMemberCount().ConfigureAwait(false);
            
             return this;
         }
