+++ src/StratisMasternodeDashboard/Services/MultiSigNodeDataService.cs
@@ -45,9 +45,7 @@ namespace Stratis.FederatedSidechains.AdminDashboard.Services
             await UpdateMultiSig().ConfigureAwait(false);
 
             // Sidechain related updates.
-            PendingPolls = await UpdatePolls().ConfigureAwait(false);
             KickFederationMememberPendingPolls = await UpdateKickFederationMemberPolls().ConfigureAwait(false);
-            FederationMemberCount = await UpdateFederationMemberCount().ConfigureAwait(false);
            
             return this;
         }
