+++ src/StratisMasternodeDashboard/Views/Shared/Partials/ProofOfAuthority.cshtml
@@ -46,6 +46,8 @@
                     <button type="button" class="btn btn-primary" onclick="submitVote()">Submit</button>
                 </form>
                 <label class="mt-3"><strong>Pending Votes :</strong></label>
+                <div class="alert alert-danger" role="alert" style="display: none;" id="loadPendingVotesError">
+                </div>         
                 <table class="table table-sm table-striped table-bordered" id="polls-table">
                     <thead>
                         <tr>
