+++ src/StratisMasternodeDashboard/Views/Shared/_Layout.cshtml
@@ -85,11 +85,11 @@
                     <a data-toggle="modal" href="#federationVotesIDGBKick"><i class="fas fa-vote-yea"></i>&nbsp; Federation Member Kicking</a>
                 </li>
                 <li class="list-inline-item">
-                    <a data-toggle="modal" href="#federationVotes"><i class="fas fa-vote-yea"></i>&nbsp; Contract Whitelisting</a>
+                    <a data-toggle="modal" href="#federationVotes" onclick="GetPendingPoll()"><i class="fas fa-vote-yea"></i>&nbsp; Contract Whitelisting</a>
                 </li>
                 @*<li class="list-inline-item ml-4">
-                    <a asp-controller="Home" asp-action="Shutdown" class="text-danger"><i class="fas fa-power-off"></i>&nbsp; Shutdown Dashboard</a>
-                    </li>*@
+                <a asp-controller="Home" asp-action="Shutdown" class="text-danger"><i class="fas fa-power-off"></i>&nbsp; Shutdown Dashboard</a>
+                </li>*@
             </ul>
         </div>
     </nav>
@@ -100,3 +100,21 @@
     @RenderSection("Scripts", false)
 </body>
 </html>
+
+<script type="text/javascript">
+    function GetPendingPoll() {
+        $.get("/sidechain-node/Getpolls")
+            .done(function (data) {
+
+            })
+            .fail(function (result) {
+                if (!!result.responseText) {
+                    $("#loadPendingVotesError").text(result.responseText);
+                    $("#loadPendingVotesError").show();
+                    return;
+                }
+                $("#loadPendingVotesError").text("Failed to load. Unknown error. Please try again and make sure nodes are running.");
+                $("#loadPendingVotesError").show();
+            });
+    }
+</script>
